Restore models to default values when meeting ends.
With this change users should be able to create the manager,
client, and roster once and reuse for multiple meeting connections.
For some reason recreating the manager, client, and roster instances
for each meeting does not work as expected. It will only work for the
first meeting and fail for all subsequent attempts. This patch acts
as a workaround. There should be a follow up patch updating
documentation for how to use properly..