Fix video mute unmute again
The previous fix involved restarting the local media stream on unmute
which caused issues with orphaning a stream when leaving a meeting.
Here we just make sure the local media manager state updates to null
when video is muted which allows it to trigger appropriate state
changes when recreating the stream on unmute.